dfras-backend/services/ai-query-service/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
sqlalchemy
httpx
certifi
urllib3
psycopg2-binary
nltk
textblob
google-generativeai
python-dotenv
transformers
torch
